feat: Implement embedding models and utility functions for text processing, including OpenAI and SentenceTransformer models
