+++ kura/base_classes/embedding.py
@@ -6,3 +6,9 @@ class BaseEmbeddingModel(ABC):
     async def embed(self, texts: list[str]) -> list[list[float]]:
         """Embed a list of texts into a list of lists of floats"""
         pass
+
+    def slug(self) -> str:
+        """Return a unique identifier for the embedding model.
+        This is used to identify the embedding model in the checkpoint manager.
+        """
+        pass
